general restructure for gin (non functional)
