+++ internal/api/routes/routes.go
@@ -1,0 +1,25 @@
+package routes
+
+import (
+	"github.com/gin-gonic/gin"
+	"github.com/acmcsufoss/api.acmcsuf.com/internal/api/services"
+	"github.com/acmcsufoss/api.acmcsuf.com/internal/api/handlers"
+)
+
+func SetupRoutes(router *gin.Engine, eventService *services.EventsService) {
+	eventsHandler := handlers.NewEventHandler(eventService)
+
+	router.GET("/events")
+	router.GET("/events/:id")
+	router.POST("/events")
+	router.POST("/events/:id")
+	router.DELETE("/events/:id")
+	// ss.Get(path, s.Resources())
+	// ss.Post(path, s.PostResources())
+	// ss.Post(path, s.BatchPostResources())
+	// ss.Get(path+"/{id}", s.Resource())
+	// ss.Post(path+"/{id}", s.PostResource())
+	// ss.Post(path+"/{id}", s.BatchPostResource())
+	// ss.Delete(path+"/{id}", s.DeleteResource())
+
+}
