--- internal/api/services/interfaces.go
@@ -1,30 +1,0 @@
-package services
-
-//"github.com/swaggest/usecase"
-
-type Service interface {
-	/*
-		Old resource service interface, might be replaced later
-
-		// Resources gets a list of paginated resource resources.
-		Resources() usecase.IOInteractor
-
-		// PostResources creates a new resource resource.
-		PostResources() usecase.IOInteractor
-
-		// BatchPostResources creates multiple new resource resources.
-		BatchPostResources() usecase.IOInteractor
-
-		// Resource gets a single resource resource.
-		Resource() usecase.IOInteractor
-
-		// PostResource updates a single resource resource.
-		PostResource() usecase.IOInteractor
-
-		// BatchPostResource updates multiple resource resources.
-		BatchPostResource() usecase.IOInteractor
-
-		// DeleteResource deletes a single resource resource.
-		DeleteResource() usecase.IOInteractor
-	*/
-}
